+++ python/pathway/xpacks/llm/_utils.py
@@ -1,0 +1,43 @@
+# Copyright © 2024 Pathway
+
+import asyncio
+import functools
+import threading
+from collections.abc import Callable
+
+
+# https://stackoverflow.com/a/75094151
+class _RunThread(threading.Thread):
+    def __init__(self, coroutine):
+        self.coroutine = coroutine
+        self.result = None
+        super().__init__()
+
+    def run(self):
+        self.result = asyncio.run(self.coroutine)
+
+
+def _run_async(coroutine):
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        loop = None
+    if loop and loop.is_running():
+        thread = _RunThread(coroutine)
+        thread.start()
+        thread.join()
+        return thread.result
+    else:
+        return asyncio.run(coroutine)
+
+
+def _coerce_sync(func: Callable) -> Callable:
+    if asyncio.iscoroutinefunction(func):
+
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            return _run_async(func(*args, **kwargs))
+
+        return wrapper
+    else:
+        return func